wip: move browser's code to lace common
