+++ apps/browser-extension-wallet/src/features/delegation/stores/createDelegationStore.ts
@@ -1,10 +1,9 @@
 /* eslint-disable no-magic-numbers */
 import create, { StateSelector } from 'zustand';
 import { Wallet } from '@lace/cardano';
-import { formatPercentages, getRandomIcon } from '@lace/common';
+import { formatPercentages, getNumberWithUnit, getRandomIcon } from '@lace/common';
 import { CardanoStakePool } from '../../../types';
 import { DelegationStore, stakePoolDetailsSelectorProps } from '../types';
-import { getNumberWithUnit } from '@src/utils/format-number';
 import { TxBuilder } from '@cardano-sdk/tx-construction';
 
 export const stakePoolDetailsSelector: StateSelector<DelegationStore, stakePoolDetailsSelectorProps> = ({
