+++ apps/browser-extension-wallet/src/utils/__tests__/format-number.test.ts
@@ -63,43 +63,6 @@ describe('format-number utils', () => {
     });
   });
 
-  describe('getNumberWithUnit', () => {
-    test('formats a number rounding up to 2 decimal places according to its unit', () => {
-      expect(formatNumber.getNumberWithUnit('10234')).toEqual({ number: '10.23', unit: 'K' });
-      expect(formatNumber.getNumberWithUnit('10235')).toEqual({ number: '10.24', unit: 'K' });
-      expect(formatNumber.getNumberWithUnit('10235000')).toEqual({ number: '10.24', unit: 'M' });
-      expect(formatNumber.getNumberWithUnit('10235000000')).toEqual({ number: '10.24', unit: 'B' });
-      expect(formatNumber.getNumberWithUnit('10235000000000')).toEqual({ number: '10.24', unit: 'T' });
-      expect(formatNumber.getNumberWithUnit('10235000000000000')).toEqual({ number: '10.24', unit: 'Q' });
-    });
-
-    test(
-      'formats a number rounding up to 2 decimal places and returns an empty string as the unit ' +
-        'when the number is less than 1000',
-      () => {
-        expect(formatNumber.getNumberWithUnit('999')).toEqual({ number: '999', unit: '' });
-        expect(formatNumber.getNumberWithUnit('999.99')).toEqual({ number: '999.99', unit: '' });
-        expect(formatNumber.getNumberWithUnit('999.991')).toEqual({ number: '999.99', unit: '' });
-        expect(formatNumber.getNumberWithUnit('999.999')).toEqual({ number: '1000', unit: '' });
-      }
-    );
-
-    test('returns the same value and no unit in case of a NaN value', () => {
-      expect(formatNumber.getNumberWithUnit('asd')).toEqual({ number: 'asd' });
-    });
-
-    test('formats negatives and decimal values', () => {
-      expect(formatNumber.getNumberWithUnit('-912180')).toEqual({ number: '-912.18', unit: 'K' });
-      expect(formatNumber.getNumberWithUnit('123452.2222')).toEqual({ number: '123.45', unit: 'K' });
-      expect(formatNumber.getNumberWithUnit('123455.5555')).toEqual({ number: '123.46', unit: 'K' });
-    });
-
-    test('removes any leading or trailing zeroes while formatting', () => {
-      expect(formatNumber.getNumberWithUnit('0000010234')).toEqual({ number: '10.23', unit: 'K' });
-      expect(formatNumber.getNumberWithUnit('1000.00000')).toEqual({ number: '1', unit: 'K' });
-    });
-  });
-
   describe('compactNumberWithUnit', () => {
     test('completes the decimal part with the correct amount of zeroes', () => {
       expect(formatNumber.compactNumberWithUnit('10')).toEqual('10.00');
