+++ packages/common/src/ui/lib/__tests__/format-number.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-magic-numbers */
-import { formatPercentages } from '../format-number';
+import { formatPercentages, getNumberWithUnit } from '../format-number';
 
 describe('Number formatters', () => {
   describe('formatPercentages', () => {
@@ -18,5 +18,42 @@ describe('Number formatters', () => {
     test('formats percentages with many decimals', () => {
       expect(formatPercentages(0.339, { decimalPlaces: 6, rounding: 'down' })).toEqual('33.900000');
     });
+  });
+});
+
+describe('getNumberWithUnit', () => {
+  test('formats a number rounding up to 2 decimal places according to its unit', () => {
+    expect(getNumberWithUnit('10234')).toEqual({ number: '10.23', unit: 'K' });
+    expect(getNumberWithUnit('10235')).toEqual({ number: '10.24', unit: 'K' });
+    expect(getNumberWithUnit('10235000')).toEqual({ number: '10.24', unit: 'M' });
+    expect(getNumberWithUnit('10235000000')).toEqual({ number: '10.24', unit: 'B' });
+    expect(getNumberWithUnit('10235000000000')).toEqual({ number: '10.24', unit: 'T' });
+    expect(getNumberWithUnit('10235000000000000')).toEqual({ number: '10.24', unit: 'Q' });
+  });
+
+  test(
+    'formats a number rounding up to 2 decimal places and returns an empty string as the unit ' +
+      'when the number is less than 1000',
+    () => {
+      expect(getNumberWithUnit('999')).toEqual({ number: '999', unit: '' });
+      expect(getNumberWithUnit('999.99')).toEqual({ number: '999.99', unit: '' });
+      expect(getNumberWithUnit('999.991')).toEqual({ number: '999.99', unit: '' });
+      expect(getNumberWithUnit('999.999')).toEqual({ number: '1000', unit: '' });
+    }
+  );
+
+  test('returns the same value and no unit in case of a NaN value', () => {
+    expect(getNumberWithUnit('asd')).toEqual({ number: 'asd' });
+  });
+
+  test('formats negatives and decimal values', () => {
+    expect(getNumberWithUnit('-912180')).toEqual({ number: '-912.18', unit: 'K' });
+    expect(getNumberWithUnit('123452.2222')).toEqual({ number: '123.45', unit: 'K' });
+    expect(getNumberWithUnit('123455.5555')).toEqual({ number: '123.46', unit: 'K' });
+  });
+
+  test('removes any leading or trailing zeroes while formatting', () => {
+    expect(getNumberWithUnit('0000010234')).toEqual({ number: '10.23', unit: 'K' });
+    expect(getNumberWithUnit('1000.00000')).toEqual({ number: '1', unit: 'K' });
   });
 });
