+++ packages/common/src/ui/lib/format-number.ts
@@ -1,5 +1,7 @@
 /* eslint-disable no-magic-numbers */
+import BigNumber from 'bignumber.js';
 import { Percent } from '@cardano-sdk/util';
+import { UnitThreshold, getNumberUnit } from './get-number-unit';
 
 export const DEFAULT_DECIMALS = 2;
 
@@ -8,6 +10,21 @@ type FormatPercentagesOptions = {
   // rounding terminology inspired by bignumber.js
   // https://mikemcl.github.io/bignumber.js/#constructor-properties
   rounding?: 'down' | 'halfUp';
+};
+
+/**
+ * Formats a numeric string to have a maximum of two decimal places and returns its corresponding unit.
+ *
+ * @param number The number string to be formatted and to get its unit
+ * @returns An object with the formatted number and its corresponding unit
+ */
+export const getNumberWithUnit = (number: string): { number: string; unit?: string } => {
+  const bigNumber = new BigNumber(number);
+  if (bigNumber.isNaN()) return { number };
+
+  const { unit, unitThreshold } = getNumberUnit(bigNumber);
+  const threshold = unitThreshold === UnitThreshold.ZERO ? 1 : unitThreshold;
+  return { number: bigNumber.div(threshold).decimalPlaces(2).toString(), unit };
 };
 
 export const formatPercentages = (
