+++ packages/common/src/ui/lib/index.ts
@@ -3,3 +3,4 @@ export * from './format-number';
 export * from './get-random-icon';
 export * from './text-formatters';
 export * from './i18n';
+export * from './get-number-unit';
